Added DOM test for loading safe #75
